+++ src/secret.c
@@ -302,7 +302,7 @@ SEXP hash_to_sexp(unsigned char *buf, size_t sz, int conv) {
 
 static SEXP secretbase_sha3_impl(const SEXP x, const SEXP bits, const SEXP convert,
                                  void (*const hash_func)(mbedtls_sha3_context *, SEXP),
-                                 const int keccak) {
+                                 const int type) {
   
   const int conv = LOGICAL(convert)[0];
   const int bt = Rf_asInteger(bits);
@@ -312,20 +312,24 @@ static SEXP secretbase_sha3_impl(const SEXP x, const SEXP bits, const SEXP conve
   unsigned char buf[sz];
   
   mbedtls_sha3_id id;
-  switch(bt) {
-  case 256:
-    id = keccak ? MBEDTLS_KECCAK_256 : MBEDTLS_SHA3_256; break;
-  case 512:
-    id = keccak ? MBEDTLS_KECCAK_512 : MBEDTLS_SHA3_512; break;
-  case 224:
-    id = keccak ? MBEDTLS_KECCAK_224 : MBEDTLS_SHA3_224; break;
-  case 384:
-    id = keccak ? MBEDTLS_KECCAK_384 : MBEDTLS_SHA3_384; break;
-  default:
-    if (keccak) Rf_error("'bits' must be 224, 256, 384 or 512");
+  if (type) {
+    switch(bt) {
+    case 256:
+      id = type > 0 ? MBEDTLS_SHA3_256 : MBEDTLS_KECCAK_256; break;
+    case 512:
+      id = type > 0 ? MBEDTLS_SHA3_512 : MBEDTLS_KECCAK_512; break;
+    case 224:
+      id = type > 0 ? MBEDTLS_SHA3_224 : MBEDTLS_KECCAK_224; break;
+    case 384:
+      id = type > 0 ? MBEDTLS_SHA3_384 : MBEDTLS_KECCAK_384; break;
+    default:
+      if (type < 0) Rf_error("'bits' must be 224, 256, 384 or 512");
+      id = MBEDTLS_SHA3_SHAKE256;
+    }
+  } else {
     id = MBEDTLS_SHA3_SHAKE256;
   }
-
+  
   mbedtls_sha3_context ctx;
   mbedtls_sha3_init(&ctx);
   mbedtls_sha3_starts(&ctx, id);
@@ -341,11 +345,23 @@ static SEXP secretbase_sha3_impl(const SEXP x, const SEXP bits, const SEXP conve
 
 SEXP secretbase_sha3(SEXP x, SEXP bits, SEXP convert) {
   
-  return secretbase_sha3_impl(x, bits, convert, hash_object, 0);
+  return secretbase_sha3_impl(x, bits, convert, hash_object, 1);
   
 }
 
 SEXP secretbase_sha3_file(SEXP x, SEXP bits, SEXP convert) {
+  
+  return secretbase_sha3_impl(x, bits, convert, hash_file, 1);
+  
+}
+
+SEXP secretbase_shake256(SEXP x, SEXP bits, SEXP convert) {
+  
+  return secretbase_sha3_impl(x, bits, convert, hash_object, 0);
+  
+}
+
+SEXP secretbase_shake256_file(SEXP x, SEXP bits, SEXP convert) {
   
   return secretbase_sha3_impl(x, bits, convert, hash_file, 0);
   
@@ -353,12 +369,12 @@ SEXP secretbase_sha3_file(SEXP x, SEXP bits, SEXP convert) {
 
 SEXP secretbase_keccak(SEXP x, SEXP bits, SEXP convert) {
   
-  return secretbase_sha3_impl(x, bits, convert, hash_object, 1);
+  return secretbase_sha3_impl(x, bits, convert, hash_object, -1);
   
 }
 
 SEXP secretbase_keccak_file(SEXP x, SEXP bits, SEXP convert) {
   
-  return secretbase_sha3_impl(x, bits, convert, hash_file, 1);
+  return secretbase_sha3_impl(x, bits, convert, hash_file, -1);
   
 }